# integrations/conversational-skills/src/api-model/conversationalSkill.ts
/**
 * Procode Endpoints for Conversational Skills
 * This document describes the following APIs that will be implemented by a watsonx Assistant client, and consumed by the watsonx Assistant team in order to enable the objectives set forth by the Conversational Skills Mission.
 *
 * OpenAPI spec version: 1.0.11
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Information about a conversational skill
 */
export interface ConversationalSkill {
  /**
   * The unique identifier of a conversational skill
   */
  id?: string;
  /**
   * The name of a conversational skill
   */
  name?: string;
  /**
   * The description of a conversational skill
   */
  description?: string;
  /**
   * The created timestamp of a conversational skill
   */
  created?: string;
  /**
   * The created timestamp of a conversational skill
   */
  modified?: string;
  /**
   * Additional metadata of a conversational skill
   */
  metadata?: any;
}