# integrations/conversational-skills/src/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Procode Endpoints for Conversational Skills
  description: |-
    This document describes the following APIs that will be implemented by a watsonx Assistant client, and consumed by the watsonx Assistant team in order to enable the objectives set forth by the Conversational Skills Mission. 
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: www.my-hosted-provider.com
...
paths:
  /providers/{provider_id}/conversational_skills:
    get:
      tags:
      - Conversational skill
      summary: Fetch a list of conversational skills
      description: "Retrieves a list of conversational skills associated to a particular provider."
      operationId: fetchSkills
      parameters:
      - name: provider_id 
        in: path
        description: Unique identifier of the provider that possesses the conversational skill. It represents the instance that is linked with the WxA instance.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: assistant_id
        in: query
        description: Assistant ID values that need to be considered for filtering
        required: true
        explode: true
        schema:
          type: string
          default: available
      - name: environment_id
        in: query
        description: Environment ID values that need to be considered for filtering
        required: true
        explode: true
        schema:
          type: string
          default: available
      responses:
        "200":
          description: Successful request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSkillsResponse'
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ConversationalSkill:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of a conversational skill
        name:
          type: string
          description: The name of a conversational skill
        description:
          type: string
          description: The description of a conversational skill
        created:
          type: string
          description: The created timestamp of a conversational skill
        modified:
          type: string
          description: The created timestamp of a conversational skill
        metadata:
          type: object
          description: Additional metadata of a conversational skill
      description: Information about a conversational skill
    ListSkillsResponse: 
      type: object
      properties: 
        conversational_skills: 
          type: array
          description: |-
            An array of conversational_skill objects. Each item in the array should be a unique conversational skill.
          items:
            $ref: '#/components/schemas/ConversationalSkill'
...
    ErrorResponse:
      required:
      - code
      - error
      properties:
        error:
          type: string
          description: General description of an error.
        errors:
          type: array
          description: Collection of specific constraint violations associated with the error.
          items:
            $ref: '#/components/schemas/ErrorDetail'
        code:
          type: integer
          description: HTTP status code for the error response.
    ErrorDetail:
      required:
      - message
      properties:
        message:
          type: string
          description: Description of a specific constraint violation.
        path:
          type: string
          description: The location of the constraint violation.
...